Use JS objects to reporesent templates instead of JSON
In order to get a reference to the Object for rendering the template, you need
some way to identify it. In the previous JSON approach, we were dependent on
using a String to encode this data. Unfortunately, the only way to deduce a
reference to a function from a String is to use `eval()` or something equally
as gross. Instead, let's just use actual objects now and figure out dynamic
template loading later.